Reorganize addresses #21
- Move conformers to conformers.py
- Move qss_importer to importers.py
- Move custom_functions to functions.py
- Move FileSystemEventHandler to events.py
- Move main to __main__.py
- Rename compile_css to compile
- Rename compile_css_and_save to compile_and_save
- Import api methods compile and compile_and_save in __init__.py
